# openapi.yaml
openapi: 3.0.3
info:
  title: eBadge ID API
  description: >
    API pública del backend principal de eBadge ID. Cubre la superficie
    completa expuesta bajo `/api` por el backend principal: autenticación
    (local y SSO/SAML), organizaciones, usuarios, credenciales, plantillas
    de diseño, contratos digitales, metas/tareas/puntajes, notificaciones,
    API keys, invitaciones, activos de organización, brand kit, uploads,
    integraciones LMS, anclaje blockchain y los tres tableros de
    analítica/desempeño.

    Cerrado como parte del hallazgo H-15 de la auditoría técnica integral
    (2026-08): antes solo documentaba el subconjunto pensado para
    integración externa (credenciales/plantillas/auth). Cada endpoint de
    abajo se extrajo leyendo directamente su archivo de rutas y su
    controlador real — método, parámetros, autenticación, cuerpo de
    petición y códigos de respuesta reflejan el código tal como está, no
    un diseño aspiracional. Las mesa de ayuda (`help_backend`) es un
    servicio separado, con su propia base URL, y no está cubierta aquí.
  version: "1.1"
servers:
  - url: https://api.ebadgeid.com/api
    description: Producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token de sesión de usuario (POST /auth/login), enviado como
        `Authorization: Bearer <token>` o, alternativamente, como la cookie
        httpOnly `ebadge_token` que el propio login también setea.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Clave de integración por organización (ver POST /keys/{organization_code}).
        Alternativa a bearerAuth para llamadas programáticas — no requiere
        una sesión de usuario humano. Sujeta a su propio límite de
        requests/minuto (requests_allowed_per_minute).
  parameters:
    limitParam:
      name: limit
      in: query
      schema: { type: integer, default: 200, maximum: 500 }
      description: Máximo de resultados. Se recorta a 500 aunque se pida más.
    skipParam:
      name: skip
      in: query
      schema: { type: integer, default: 0 }
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        error: { type: string, description: "Algunos controladores usan `error` en vez de `message` para el mismo propósito -- ambos son texto plano para mostrar o loguear, nunca un código machine-readable." }
        details: { type: string }
    Organization:
      type: object
      properties:
        _id: { type: string }
        organization_code: { type: string, example: "ORG-A1B2C3D4" }
        name: { type: string }
        city: { type: string }
        state: { type: string }
        country: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        status: { type: string, enum: [ACTIVE, TRIAL, DELETED] }
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: >
            Set cuando status es DELETED (hallazgo H-22: DELETE
            /organizations/{id} es un borrado lógico, no físico -- la
            organización y todos sus datos se conservan para
            auditoría/retención; una purga permanente controlada es una
            decisión separada, deliberadamente no implementada).
        plan: { type: string, enum: [Free, Basic, Premium, Enterprise] }
        users: { type: integer }
        logo: { type: string }
        signature: { type: string }
        require_bulk_approval:
          type: boolean
          description: Si es true, POST /credentials/bulk-issue queda pendiente de aprobación en vez de emitir de inmediato.
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200':
          description: >
            Sesión iniciada. El token se devuelve en el body Y se setea como
            cookie httpOnly (`ebadge_token`) — ver AUDIT_FIXES.md.
          content:
            application/json:
              schema:
                type: object
                properties:
                  username: { type: string }
                  role: { type: string, enum: [admin, user, platform_admin] }
                  organization_code: { type: string }
        '400':
          description: >
            Credenciales inválidas, cuenta pendiente de activación, o la
            organización del usuario fue eliminada (H-22) -- las tres
            devuelven el mismo mensaje genérico para no filtrar cuál es el
            caso real.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Crear credenciales de acceso para un perfil ya existente (legacy)
      description: >
        Requiere que el perfil (Users) ya exista bajo la organización del
        admin autenticado. El flujo normal para altas nuevas es POST
        /users/self-signup (invitación) o POST /organizations/self-signup.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password, minLength: 12 }
      responses:
        '201': { description: Usuario registrado }
        '400': { description: "Faltan campos, contraseña muy corta, o el usuario ya existe" }
        '403': { description: Requiere rol admin }
        '404': { description: No existe un perfil (Users) para ese username en esta organización }
...
  /organizations:
    post:
      tags: [Organizations]
      summary: Crear una organización (uso administrativo directo, no autoservicio)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, admin_email, admin_phone, admin_first_name, admin_last_name, city, state, country]
              properties:
                name: { type: string }
                admin_email: { type: string, format: email }
                admin_phone: { type: string }
                admin_first_name: { type: string }
                admin_last_name: { type: string }
                city: { type: string }
                state: { type: string }
                country: { type: string }
      responses:
        '201': { description: "Creada, con email de activación para el admin" }
        '400': { description: Campos inválidos }
        '403': { description: Requiere rol platform_admin }
        '409': { description: Ya existe un admin con ese email }
    get:
      tags: [Organizations]
      summary: Listar organizaciones
      description: >
        Un platform_admin ve todas (excluyendo las eliminadas -- H-22 --
        salvo que pase `include_deleted=true`); cualquier otro rol admin
        solo ve la propia.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/limitParam' }
        - { $ref: '#/components/parameters/skipParam' }
        - name: include_deleted
          in: query
          schema: { type: boolean, default: false }
          description: Solo tiene efecto para platform_admin (H-22).
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Organization' } } } } }
...